backend/requirements/base.txt
# Django core
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-environ>=0.11,<1.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24.0

# PostGIS
psycopg[binary]>=3.1,<4.0

# Background Tasks (Django-Q with PostgreSQL ORM broker)
django-q2>=1.6,<2.0

# GIS/Raster processing
rasterio>=1.3,<2.0
rasterstats>=0.19,<1.0
pyproj>=3.6,<4.0
shapely>=2.0,<3.0
geopandas>=0.14,<1.0
fiona>=1.9,<2.0

# Bias correction
mgwr>=2.1,<3.0
scikit-learn>=1.3,<2.0
numpy>=1.24,<2.0
pandas>=2.0,<3.0
scipy>=1.11,<2.0
libpysal>=4.9,<5.0

# Reports - ReportLab instead of WeasyPrint (no GTK dependencies)
reportlab>=4.0,<5.0
matplotlib>=3.8,<4.0
Pillow>=10.0,<11.0

# HTTP clients
requests>=2.31,<3.0
httpx>=0.25,<1.0
openaq>=0.4,<1.0

# Auth
djangorestframework-simplejwt>=5.3,<6.0

# API Documentation
drf-yasg>=1.21,<2.0
# GIS Serializers and fields
djangorestframework-gis>=1.0,<2.0

# Utils
python-dateutil>=2.8,<3.0
pytz>=2023.3
